Separate files into folders, add error handling to json_loader method
