--- books.json
@@ -1,1 +1,0 @@
-[{"json_class":"book","title":"Tolkien","author":"Tolkien","id":1858}]
