+++ classes/corrector.rb
