+++ data/books.json
@@ -1,0 +1,1 @@
+[{"json_class":"book","title":"Tolkien","author":"Tolkien","id":1858},{"json_class":"book","title":"Lord of the rings 2","author":"Tolkien","id":1896}]
