+++ data/rentals.json
@@ -1,0 +1,1 @@
+null
