+++ loader.rb
@@ -2,5 +2,7 @@ class Loader
   def self.json_loader(path)
     file = File.read(path)
     JSON.parse(file)
+  rescue StandardError
+    puts "Error, #{path} file not found"
   end
 end
