--- persons.json
@@ -1,1 +1,0 @@
-[{"json_class":"Student","age":34,"parent_permission":true,"name":"Carlos","id":453},{"json_class":"teacher","specialization":"Vet","age":55,"name":"Julian","id":75}]
