--- rentals.json
@@ -1,1 +1,0 @@
-[{"json_class":"rental","date":"2000/12/05","book":{"json_class":"book","title":"Tolkien","author":"Tolkien","id":1858},"person":{"json_class":"Student","age":34,"parent_permission":true,"name":"Carlos","id":453}},{"json_class":"rental","date":"1999/03/08","book":{"json_class":"book","title":"Tolkien","author":"Tolkien","id":1858},"person":{"json_class":"Student","age":34,"parent_permission":true,"name":"Carlos","id":453}},{"json_class":"rental","date":"1800/05/03","book":{"json_class":"book","title":"Tolkien","author":"Tolkien","id":1858},"person":{"json_class":"teacher","specialization":"Vet","age":55,"name":"Julian","id":75}}]
